:art: Re-added Katalog( Kehapus)

+++ body.cpp
@@ -857,58 +857,53 @@ void belanja(Gudang gudang)
     int jumBar, returnValue, harga;
     addrBar beliBar;
     addrBel awal = NULL, akhir = NULL, curr;
-
-    do
+    if(katalog(gudang.root))
     {
-        printf("Beli: ");
-        scanf(" %[^\n]s", &beli);
-        printf("Jumlah: ");
-        scanf("%d", &jumBar);
-
-        while (!cekKategori(gudang.root, beli))
+        do
         {
-            printf("%s tidak ada\n", beli);
             printf("Beli: ");
             scanf(" %[^\n]s", &beli);
             printf("Jumlah: ");
             scanf("%d", &jumBar);
-        }
 
-        beliBar = searchGudang(gudang.root, beli);
-        if (beliBar != NULL)
+            while (!cekKategori(gudang.root, beli))
+            {
+                printf("%s tidak ada\n", beli);
+                printf("Beli: ");
+                scanf(" %[^\n]s", &beli);
+                printf("Jumlah: ");
+                scanf("%d", &jumBar);
+            }
+
+            beliBar = searchGudang(gudang.root, beli);
+            if (beliBar != NULL)
+            {
+                insertBarBel(gudang, beliBar, &awal, &akhir, jumBar);
+            }
+            else
+            {
+                puts("Barang tidak ada");
+                break;
+            }
+            printf("Ada lagi yang dibeli? (y/n): ");
+            scanf(" %c", &lanjut);
+
+        } while (lanjut == 'y' || lanjut == 'Y');
+        system("cls");
+        puts("Struk Hafiz Market");
+        returnValue = tampilBarBel(awal, akhir);
+        system("pause");
+        system("cls");
+        transaksi(returnValue, gudang, awal);
+        // ? Proses delete pembelian
+        curr = awal;
+        while (curr->nextBar != NULL)
         {
-            insertBarBel(gudang, beliBar, &awal, &akhir, jumBar);
+            curr = curr->nextBar;
+            deleteBarBel(&awal, akhir);
         }
-        else
-        {
-            puts("Barang tidak ada");
-            break;
-        }
-        printf("Ada lagi yang dibeli? (y/n): ");
-        scanf(" %c", &lanjut);
-
-    } while (lanjut == 'y' || lanjut == 'Y');
-    system("cls");
-    puts("Struk Hafiz Market");
-    returnValue = tampilBarBel(awal, akhir);
-    system("pause");
-    system("cls");
-    puts("Pembayaran");
-    printf("Total harga          : Rp %d\n", returnValue);
-    printf("Masukkan jumlah uang : Rp ");
-    scanf(" %d", &harga);
-    printf("Kembalian            : Rp %d\n", harga - returnValue);
-    puts("Terima kasih sudah belanja!");
-    kurangistock(gudang, awal);
-    system("pause");
-
-    // ? Proses delete pembelian
-    curr = awal;
-    while (curr->nextBar != NULL)
-    {
-        curr = curr->nextBar;
-        deleteBarBel(&awal, akhir);
     }
+    
 }
 
 // ? ================================================== Modul untuk menu gudang ==============================================================
